% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_melange_dist.R
\name{mean.melange_dist}
\alias{mean.melange_dist}
\title{Moyenne arithmétique}
\usage{
\method{mean}{melange_dist}(x, ...)
}
\arguments{
\item{x}{objet de classe \code{melange_dist}.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Moyenne des réalisations.
}
\description{
Moyenne des observations issus d'un objet de classe \code{melange_dist}.
}
\examples{
M <- melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
mean(M)
N <- melange_normal(c(1,.2),c(.1,.8))
mean(N)


}
