% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melange_dist.R
\name{melange_dist}
\alias{melange_dist}
\title{Mélange de distributions}
\usage{
melange_dist(x = list(), prob = vector(), ...)
}
\arguments{
\item{x}{liste de lois. Chacune doit être une fonction d’un seul argument,\code{n},
permettant de générer \code{n} réalisations d’une variable aléatoire}

\item{prob}{poids du mélange.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Liste des lois à mélanger avec leurs poids de mélanges respectives normalisés et la classe \code{melange_dist}.
}
\description{
Classe \code{S3} d'un objet de mélange de lois.
}
\examples{
melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
melange_dist(list(\(n) rnorm(n, 10, 5), \(n) rnorm(n, 3, 2)), c(3, 4))
melange_dist(list(\(n) rexp(n, 10), \(n) rexp(n, 20)), c(3, 4), "Mélange de deux exponentielles")

}
