% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melange_normal.R
\name{melange_normal}
\alias{melange_normal}
\title{Mélange de lois normales}
\usage{
melange_normal(moy = vector(), ect = vector(), prob = NULL, ...)
}
\arguments{
\item{moy}{vecteur de moyennes.}

\item{ect}{vecteur d'écart-type.}

\item{prob}{vecteur de poids du mélange.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Vecteurs de moyenne, d'écart-type et poids normalisés des lois normales à mélanger
 et la classe \code{melange_normal}.
}
\description{
Classe  \code{S3} d'un objet de mélange de lois normales. Les méthodes de la classe \code{melange_dist}
 sont aussi applicables pour cette classe.
}
\examples{
melange_normal(c(1,1),c(.1,.8))
melange_normal( c(3, 4),c(2,2),prob= c(.2,.6))

}
