% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixLaw-package.R
\docType{package}
\name{MixLaw-package}
\alias{MixLaw}
\alias{MixLaw-package}
\title{MixLaw: Génération de mélange de lois}
\description{
Ce package permet d'implémenter la classe représentant un mélange de distribution. Il permet de générer des observations issues d'un mélange de lois, de fournir une représentation de la densité correspondant aux réalisations, de calculer la moyenne et le quantile des observations.
}
\author{
\strong{Maintainer}: Komi Ayi \email{ayi.komi_roger@courrier.uqam.ca}

}
\keyword{internal}
