% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_melange_dist.R
\name{plot.melange_dist}
\alias{plot.melange_dist}
\title{Graphique}
\usage{
\method{plot}{melange_dist}(
  x,
  n = 100L,
  type = "l",
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  color = "black",
  ...
)
}
\arguments{
\item{x}{objet de classe \code{melange_dist}.}

\item{n}{nombre d'observations tracées.}

\item{type}{caractère pour désigner le type de tracer}

\item{xlim}{les limites x (x1, x2) du tracé.}

\item{ylim}{les limites y du tracé.}

\item{main}{titre du tracé, see also \code{\link{plot}}.}

\item{xlab}{label de l'axe des x.}

\item{ylab}{label de l'axe des y.}

\item{color}{couleur du tracé.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Graphique de la densité du mélange de lois.
}
\description{
Affiche la densité des réalisations d'un objet de classe \code{melange_dist}.
}
\examples{
M <- melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
plot(M, color="red")
N <- melange_normal(c(1,.2),c(1,.8))
plot(N, main="Mélange de lois normales")

}
