% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_melange_dist.R
\name{print.melange_dist}
\alias{print.melange_dist}
\title{Affichage informations}
\usage{
\method{print}{melange_dist}(x, ...)
}
\arguments{
\item{x}{objet de classe \code{melange_dist}.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Les informations sur l'instance sur laquelle elle est appliquée.
}
\description{
Affiche un tableau qui renseigne sur les informations d'une instance de classe \code{melange_dist}.
}
\examples{
M <- melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
print(M)
N <- melange_normal(c(1,.2,4),c(.1,.8,8))
print(N)

}
