% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_melange_dist.R
\name{quantile.melange_dist}
\alias{quantile.melange_dist}
\title{Quantile}
\usage{
\method{quantile}{melange_dist}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{objet de classe \code{melange_dist}.}

\item{probs}{vecteur numérique des probabilités.}

\item{\dots}{arguments supplémentaires.}
}
\value{
Quantile des réalisations.
}
\description{
Calcule le quantile des observations issus d'un objet de classe \code{melange_dist}.
}
\examples{
M <- melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
quantile(M)
N <- melange_normal(c(1,.2),c(.1,.8))
quantile(N)

}
