% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_melange_dist.R
\name{rand}
\alias{rand}
\alias{rand.melange_dist}
\title{Réalisations de mélange de lois}
\usage{
rand(x, ...)

\method{rand}{melange_dist}(x, n = 100L, ...)
}
\arguments{
\item{x}{objet de classe \code{melange_dist}.}

\item{\dots}{arguments supplémentaires.}

\item{n}{nombre d'observations.}
}
\value{
\code{n} réalisations du mélange de lois.
}
\description{
Génere des réalisations de lois d'un objet de classe \code{melange_dist}.
}
\examples{
M <- melange_dist(list(rnorm, rexp, \(n) rgamma(n, 10)), c(0.1, 0.2, 0.7))
rand(M)

}
